Update from task f2d17e2b-a524-4178-bd85-8c70d3a83e15

+++ diffusion_utils.py
@@ -156,7 +156,7 @@ def sample_diffusion(
             
             # Add variance (optional, can be deterministic)
             if variance > 0:
-                x += torch.randn_like(x) * torch.sqrt(torch.tensor(variance, device=device))
+                x += torch.randn_like(x) * torch.sqrt(torch.tensor(variance, device=device, dtype=torch.float32))
     
     return torch.clamp(x, -1, 1)
